implement review feedback and close all minor remarks
